✨ feat: 사이드바 기능, css 구현완료
