+++ src/App.css
@@ -2,27 +2,37 @@
   margin: 0;
   padding: 0;
   text-decoration: none;
+  font-family: "Noto Sans KR", sans-serif !important;
+  color: #1d1d1d;
 }
 .container {
   display: flex;
+  background: #eee;
 }
 main {
   width: 100%;
   padding: 20px;
 }
 .sidebar {
-  background: #000;
-  color: #fff;
+  background: #fff;
+  color: #1d1d1d;
   height: 100vh;
-  width: 300px;
+  width: 220px;
+  transition: all 0.5s;
 }
 .top_section {
   display: flex;
+  min-width: 24px;
   align-items: center;
-  padding: 20px 15px;
+  padding: 66px 23px;
+  transition: all 0.5s;
 }
 .logo {
-  font-size: 30px;
+  color: #32b784;
+  font-weight: 900;
+  font-size: 20px;
+  overflow: hidden;
+  transition: all 0.5s;
 }
 .bars {
   display: flex;
@@ -31,21 +41,26 @@ main {
 }
 .link {
   display: flex;
-  color: #fff;
-  padding: 10px 15px;
-  gap: 15px;
-  transition: all 0.5;
+  overflow: hidden;
+  color: #1d1d1d;
+  padding: 10px 13px;
+  gap: 13px;
+  align-items: center;
+  border-bottom: 0.5px solid #fff;
 }
 .link:hover {
-  background: lightskyblue;
-  color: #000;
-  transition: all 0.5s;
+  border-bottom: 0.5px solid #32b784;
+  color: #1d1d1d;
 }
 .active {
-  background: lightskyblue;
-  color: #000;
+  border-bottom: 0.5px solid #32b784;
+  color: #1d1d1d;
+  transition: all 0.5s;
 }
-.icon,
+
 .link_text {
-  font-size: 20px;
+  font-size: 14px;
+  font-weight: 400;
+  transition: all 0.5s;
+  white-space: nowrap;
 }
